// web-application/src/types-openapi/api/models/CreateStockOutRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CreateStockOutItem } from './CreateStockOutItem';
import {
    CreateStockOutItemFromJSON,
    CreateStockOutItemFromJSONTyped,
    CreateStockOutItemToJSON,
    CreateStockOutItemToJSONTyped,
} from './CreateStockOutItem';

/**
 * 
 * @export
 * @interface CreateStockOutRequest
 */
export interface CreateStockOutRequest {
    /**
     * 
     * @type {Date}
     * @memberof CreateStockOutRequest
     */
    outTime?: Date;
    /**
     * 
     * @type {number}
     * @memberof CreateStockOutRequest
     */
    supplierId?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateStockOutRequest
     */
    outBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateStockOutRequest
     */
    note?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CreateStockOutRequest
     */
    complete?: boolean;
    /**
     * 
     * @type {Array<CreateStockOutItem>}
     * @memberof CreateStockOutRequest
     */
    items?: Array<CreateStockOutItem> | null;
}

/**
 * Check if a given object implements the CreateStockOutRequest interface.
 */
export function instanceOfCreateStockOutRequest(value: object): value is CreateStockOutRequest {
    return true;
}

export function CreateStockOutRequestFromJSON(json: any): CreateStockOutRequest {
    return CreateStockOutRequestFromJSONTyped(json, false);
}

export function CreateStockOutRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateStockOutRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'outTime': json['outTime'] == null ? undefined : (new Date(json['outTime'])),
        'supplierId': json['supplierId'] == null ? undefined : json['supplierId'],
        'outBy': json['outBy'] == null ? undefined : json['outBy'],
        'note': json['note'] == null ? undefined : json['note'],
        'complete': json['complete'] == null ? undefined : json['complete'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(CreateStockOutItemFromJSON)),
    };
}

export function CreateStockOutRequestToJSON(json: any): CreateStockOutRequest {
    return CreateStockOutRequestToJSONTyped(json, false);
}

export function CreateStockOutRequestToJSONTyped(value?: CreateStockOutRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'outTime': value['outTime'] == null ? undefined : ((value['outTime']).toISOString()),
        'supplierId': value['supplierId'],
        'outBy': value['outBy'],
        'note': value['note'],
        'complete': value['complete'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(CreateStockOutItemToJSON)),
    };
}